server/mood_api/requirements.txt
asgiref==3.7.2
click==8.1.6
Django==4.2.4
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
isoweek==1.3.3
joblib==1.3.1
nltk==3.8.1
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
textblob==0.17.1
TimeConvert==3.0.13
tqdm==4.65.0
typing_extensions==4.7.1
tzlocal==5.0.1
xlwt==1.3.0
